// frontend/src/lib/api/model/matchScoutingRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: v2.0.2
 */
import type { MatchScoutingRequestFields } from './matchScoutingRequestFields';

export interface MatchScoutingRequest {
  submission_uuid: string;
  fields: MatchScoutingRequestFields;
  user_uuid: string;
  year: number;
  team_number: number;
  match_number: number;
  match_type: string;
  event_code: string;
  event_name: string;
  event_type: string;
  event_city: string;
  event_country: string;
  event_start_date: string;
  event_end_date: string;
}