// frontend/src/lib/api/model/getPitsForSeasonRequest.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: v2.0.0-alpha.6
 */

export interface GetPitsForSeasonRequest {
  season_uuid: string;
  event_code: string;
  event_name: string;
  event_type: string;
  event_city: string;
  event_country: string;
  event_start_date: string;
  event_end_date: string;
}